RP-161 검색 로직 리펙토링 및 버그 수정

+++ src/main/java/com/cherrypick/backend/domain/deal/application/dto/response/DealSearchResponseDTO.java
@@ -22,7 +22,7 @@ public class DealSearchResponseDTO {
     private String nickname;
     private Long badgeId;
     private String createdAt;
-    private int heat;
+    private Double heat;
     private int totalLikes;
     private int totalComments;
     private boolean isSoldout;
